+++ R/vars-select.R
@@ -163,8 +163,7 @@ vars_select <- function(.vars, ...,
   incl <- inds_combine(.vars, ind_list)
 
   # Returned names must be unique
-  nms <- vctrs::vec_as_names(names(incl), repair = "unique")
-  sel <- set_names(.vars[incl], nms)
+  sel <- set_names(.vars[incl], names(incl))
 
   # Include/.exclude specified variables
   sel <- c(setdiff2(.include, sel), sel)